Move Moveable's from_state method up to State
