+++ stonesoup/movable/movable.py
@@ -2,7 +2,7 @@
 import datetime
 from abc import abstractmethod, ABC
 from functools import lru_cache
-from typing import Sequence, Tuple, Any
+from typing import Sequence, Tuple, MutableSequence, Optional
 
 import numpy as np
 from math import cos, sin
@@ -16,7 +16,7 @@ from stonesoup.types.state import State, StateMutableSequence
 
 
 class Movable(StateMutableSequence, ABC):
-    states: Sequence[State] = Property(
+    states: MutableSequence[State] = Property(
         doc="A list of States which enables the platform's history to be "
             "accessed in simulators and for plotting. Initiated as a "
             "state, for a static platform, this would usually contain its "
@@ -27,7 +27,7 @@ class Movable(StateMutableSequence, ABC):
         doc="Mapping between platform position and state vector. For a "
             "position-only 3d platform this might be ``[0, 1, 2]``. For a "
             "position and velocity platform: ``[0, 2, 4]``")
-    velocity_mapping: Sequence[int] = Property(
+    velocity_mapping: Optional[Sequence[int]] = Property(
         default=None,
         doc="Mapping between platform velocity and state dims. If not "
             "set, it will default to ``[m+1 for m in position_mapping]``")
@@ -46,45 +46,6 @@ class Movable(StateMutableSequence, ABC):
             self.velocity_mapping = [p + 1 for p in self.position_mapping]
         if not self.states:
             raise ValueError('States must not be empty: it must contain least one state.')
-
-    @staticmethod
-    def from_state(state: State, *args: Any, **kwargs: Any) -> State:
-        """Class utility function to create a new state. The type and properties of this new state
-        are defined by `state`. Specific properties can be overwritten via `args` and `kwargs`.
-
-        This method was introduced to combat the need for deep copies in platform `move` methods,
-        and to allow platforms to use different state types, such as :class:`~.GroundTruthState`,
-        whereby the state metadata will be carried forward at each call to :meth:`move`.
-        The process here is similar to that of the :class:`~.Prediction` and :class:`~.Update`
-        `from_state` methods.
-
-        Parameters
-        ----------
-        state: State
-            :class:`~.State` to use existing properties from, and identify new state-type from.
-        \\*args: Sequence
-            Arguments to pass to newly created state, replacing those with same name in `state`.
-        \\*\\*kwargs: Mapping
-            New property names and associate value for use in newly created state, replacing those
-            on the `state` parameter.
-        """
-        # Handle being initialised with state sequence
-        if isinstance(state, StateMutableSequence):
-            state = state.state
-
-        state_type = type(state)
-
-        args_property_names = {
-            name for n, name in enumerate(state_type.properties) if n < len(args)}
-
-        new_kwargs = {
-            name: getattr(state, name)
-            for name in state_type.properties.keys()
-            if name not in args_property_names}
-
-        new_kwargs.update(kwargs)
-
-        return state_type(*args, **new_kwargs)
 
     @property
     def position(self) -> StateVector:
@@ -259,7 +220,7 @@ class FixedMovable(Movable):
 
     def move(self, timestamp: datetime.datetime, **kwargs) -> None:
         """For a fixed platform this method has no effect other than to update the timestamp."""
-        new_state = self.from_state(self.state, timestamp=timestamp)
+        new_state = State.from_state(self.state, timestamp=timestamp)
         self.states.append(new_state)
 
 
@@ -379,7 +340,7 @@ class MovingMovable(Movable):
                                                       timestamp=timestamp,
                                                       time_interval=time_interval,
                                                       **kwargs)
-        new_state = self.from_state(self.state, state_vector=state_vector, timestamp=timestamp)
+        new_state = State.from_state(self.state, state_vector=state_vector, timestamp=timestamp)
         self.states.append(new_state)
 
 
@@ -451,9 +412,9 @@ class MultiTransitionMovable(MovingMovable):
                     time_interval=self.current_interval,
                     **kwargs
                 )
-                temp_state = self.from_state(self.state,
-                                             state_vector=temp_state_vector,
-                                             timestamp=timestamp)
+                temp_state = State.from_state(self.state,
+                                              state_vector=temp_state_vector,
+                                              timestamp=timestamp)
 
                 time_interval -= self.current_interval
                 self.transition_index = (self.transition_index + 1) % len(self.transition_models)
@@ -466,9 +427,9 @@ class MultiTransitionMovable(MovingMovable):
                     time_interval=time_interval,
                     **kwargs
                 )
-                temp_state = self.from_state(self.state,
-                                             state_vector=temp_state_vector,
-                                             timestamp=timestamp)
+                temp_state = State.from_state(self.state,
+                                              state_vector=temp_state_vector,
+                                              timestamp=timestamp)
                 self.current_interval -= time_interval
                 break
         self.states.append(temp_state)
